Update of the Aries Embedded board support to Yocto 2.6 (thud)
The board supports now uses:

  - U-Boot 2018.11
  - Linux 4.18.21
  - GCC 8.2
  - GLIBC 2.28
  - Many updated target packages

For the MCVEVK, the kernel branch "v4.18/standard/intel-socfpga"
is used with the adaptions from Intel's SOCFPGA Linux tree at
"https://github.com/altera-opensource/linux-socfpga".

For the M28EVK, the NAND driver in Linux 4.18 is currently broken.
Therefore, Linux 4.14 is selected for that board. The issue has
been reported to the MTD mailing list.

Signed-off-by: Wolfgang Grandegger <wg@aries-embedded.de>