Update of the Aries Embedded board support to Yocto 3.1 (dunfell)
The board supports now use:

- U-Boot v2020.01 for MCVEVK and M28EVK
  U-Boot v2018.11 for M53EVK and MA5D4EVK
- Linux Yocto 5.4.69 for M28EVK, M53EVK and MA5D4EVK
- Linux SOCFPGA 4.14.130 and 5.4.64 LTS for MCVEVK
- GCC 9.3.0
- GLIBC 2.31
- Many updated target packages

U-Boot for the MA5D4EVK and M53EVK has not been updated, because of
trouble with the size of the primary boot loader and the broken NAND
flash support, respectively.

The support for the MCVEVP and MCVEVK is now based on the official
OpenEmbedded/Yocto BSP layer for the Altera SoCFPGA platforms at
https://github.com/kraj/meta-altera.git. It uses the SOCFPGA Linux
from https://github.com/altera-opensource/linux-socfpga.

Signed-off-by: Wolfgang Grandegger <wg@aries-embedded.de>